jwtservice/jwtauthproject/requirements.txt
djangorestframework
requests
python-dotenv
djangorestframework-simplejwt
psycopg[binary,pool]
pyotp
gunicorn

#cors
django-cors-headers